+++ CHANGELOG.md
@@ -7,7 +7,11 @@
 ### Added
 - Icons matching New UI
 - Annotating unused MTL file references in OBJ files
-- Annotating unused materials in OBJ files
+- Annotating duplicated MTL file references in OBJ files
+- Annotating unused materials in MTL files
+- Import optimizer for OBJ files
+- Quick fix for optimising MTL file references in OBJ files
+- Quick fix for removing unused materials in MTL files
 
 ### Changed
 - Use split editor (text and preview) provided by IntelliJ platform